merge-upstream: rewrite the theme system
