+++ src/components/menu-bar/settings-menu.css
@@ -1,3 +1,5 @@
+@import "../../css/colors.css";
+
 .icon {
     width: 1.5rem;
 }
@@ -64,4 +66,14 @@
     .dropdown-label {
         display: none;
     }
+}
+
+.accent-icon-outer {
+    width: 24px;
+    height: 24px;
+    border: 3px solid $menu-bar-foreground;
+    border-radius: 100%;
+    box-sizing: border-box;
+    transform-origin: center;
+    transform: scale(85%);
 }
