+++ src/components/menu-bar/settings-menu.jsx
@@ -5,8 +5,10 @@ import {FormattedMessage} from 'react-intl';
 
 import LanguageMenu from './language-menu.jsx';
 import MenuBarMenu from './menu-bar-menu.jsx';
-import ThemeMenu from './theme-menu.jsx';
 import {MenuSection} from '../menu/menu.jsx';
+import TWAccentThemeMenu from './tw-theme-accent.jsx';
+import TWGuiThemeMenu from './tw-theme-gui.jsx';
+import TWBlocksThemeMenu from './tw-theme-blocks.jsx';
 
 import menuBarStyles from './menu-bar.css';
 import styles from './settings-menu.css';
@@ -55,7 +57,13 @@ const SettingsMenu = ({
         >
             <MenuSection>
                 {canChangeLanguage && <LanguageMenu onRequestCloseSettings={onRequestClose} />}
-                {canChangeTheme && <ThemeMenu onRequestCloseSettings={onRequestClose} />}
+                {canChangeTheme && (
+                    <React.Fragment>
+                        <TWGuiThemeMenu />
+                        <TWBlocksThemeMenu />
+                        <TWAccentThemeMenu />
+                    </React.Fragment>
+                )}
             </MenuSection>
         </MenuBarMenu>
     </div>
