+++ src/containers/gui.jsx
@@ -40,7 +40,7 @@ import systemPreferencesHOC from '../lib/system-preferences-hoc.jsx';
 
 import GUIComponent from '../components/gui/gui.jsx';
 import {setIsScratchDesktop} from '../lib/isScratchDesktop.js';
-import TWThemeApplierHOC from './tw-theme-applier-hoc.jsx';
+import TWThemeManagerHOC from './tw-theme-manager-hoc.jsx';
 
 const {RequestMetadata, setMetadata, unsetMetadata} = storage.scratchFetch;
 
@@ -201,7 +201,7 @@ const ConnectedGUI = injectIntl(connect(
 const WrappedGui = compose(
     LocalizationHOC,
     ErrorBoundaryHOC('Top Level App'),
-    TWThemeApplierHOC, // componentDidUpdate() needs to run very early for icons to update immediately
+    TWThemeManagerHOC, // componentDidUpdate() needs to run very early for icons to update immediately
     FontLoaderHOC,
     // QueryParserHOC, // tw: HOC is unused
     ProjectFetcherHOC,
