+++ src/lib/themes/accent/blue.js
@@ -1,0 +1,13 @@
+const guiColors = {
+    'looks-secondary': 'hsla(215, 100%, 65%, 1)',
+    'looks-transparent': 'hsla(215, 100%, 65%, 0.35)',
+    'looks-light-transparent': 'hsla(215, 100%, 65%, 0.15)',
+    'looks-secondary-dark': 'hsla(215, 60%, 50%, 1)',
+};
+
+const blockColors = {};
+
+export {
+    guiColors,
+    blockColors
+};
