+++ src/lib/themes/blocks/dark.js
@@ -78,52 +78,7 @@ const blockColors = {
 
 const extensions = {};
 
-const guiColors = {
-    'ui-primary': '#111111',
-    'ui-secondary': '#1e1e1e',
-    'ui-tertiary': '#2e2e2e',
-
-    'ui-modal-overlay': '#333333aa',
-    'ui-modal-background': '#111111',
-    'ui-modal-foreground': '#eeeeee',
-    'ui-modal-header-background': '#333333',
-    'ui-modal-header-foreground': '#ffffff',
-
-    'ui-white': '#1e1e1e',
-
-    'ui-black-transparent': '#ffffff26',
-
-    'text-primary': '#eeeeee',
-
-    'menu-bar-background': '#333333',
-
-    'assets-background': '#111111',
-
-    'input-background': '#1e1e1e',
-
-    'popover-background': '#333333',
-
-    'badge-background': '#16202c',
-    'badge-border': '#203652',
-
-    'fullscreen-background': '#111111',
-    'fullscreen-accent': '#111111',
-
-    'page-background': '#111111',
-    'page-foreground': '#eeeeee',
-
-    'link-color': '#44aaff',
-
-    'filter-icon-black': 'invert(100%)',
-    'filter-icon-gray': 'grayscale(100%) brightness(1.7)',
-    'filter-icon-white': 'brightness(0) invert(100%)',
-
-    'paint-popover-background': '#333333',
-    'paint-filter-icon-gray': 'brightness(1.7)'
-};
-
 export {
     blockColors,
-    extensions,
-    guiColors
+    extensions
 };
