--- src/lib/themes/dark/__mocks__/index.js
@@ -1,22 +1,0 @@
-const blockColors = {
-    motion: {
-        primary: '#AAAAAA'
-    },
-    pen: {
-        primary: '#FFFFFF',
-        secondary: '#EEEEEE',
-        tertiary: '#DDDDDD'
-    },
-    text: '#BBBBBB'
-};
-
-const extensions = {
-    pen: {
-        blockIconURI: 'darkPenIcon'
-    }
-};
-
-export {
-    blockColors,
-    extensions
-};
