--- src/lib/themes/default/__mocks__/index.js
@@ -1,18 +1,0 @@
-const blockColors = {
-    motion: {
-        primary: '#111111',
-        secondary: '#222222',
-        tertiary: '#333333'
-    },
-    pen: {
-        primary: '#121212',
-        secondary: '#232323',
-        tertiary: '#343434'
-    },
-    text: '#444444',
-    workspace: '#555555'
-};
-
-export {
-    blockColors
-};
