--- src/lib/themes/tw/index.js
@@ -1,69 +1,0 @@
-const lightGuiColors = {
-    'ui-modal-overlay': '#ff4c4ce6',
-    'ui-modal-header-background': '#ff4c4c',
-
-    'motion-primary': '#ff4c4c',
-    'motion-tertiary': '#cc3333',
-
-    'looks-secondary': '#ff4c4c',
-    'looks-transparent': '#ff4d4d59',
-    'looks-light-transparent': '#ff4d4d26',
-    'looks-secondary-dark': 'hsla(0, 42%, 51%, 1)',
-
-    'extensions-primary': 'hsla(10, 85%, 65%, 1)',
-    'extensions-tertiary': 'hsla(10, 85%, 40%, 1)',
-    'extensions-transparent': 'hsla(10, 85%, 65%, 0.35)',
-    'extensions-light': 'hsla(10, 57%, 85%, 1)',
-
-    'drop-highlight': '#ff8c8c',
-
-    'menu-bar-background': '#ff4c4c'
-};
-
-const darkBlockColors = {
-    insertionMarker: '#cccccc',
-    workspace: '#1e1e1e',
-    toolboxSelected: '#1e1e1e',
-    toolboxText: '#cccccc',
-    toolbox: '#111111',
-    flyout: '#111111',
-    scrollbar: '#666666',
-    contextMenuBackground: '#111111',
-    contextMenuBorder: '#ffffff26',
-    contextMenuForeground: '#eeeeee',
-    contextMenuActiveBackground: '#2e2e2e',
-    contextMenuDisabledForeground: '#666666',
-    flyoutLabelColor: '#cccccc',
-    checkboxInactiveBackground: '#222222',
-    checkboxInactiveBorder: '#c8c8c8',
-    checkboxActiveBackground: '#ff4c4c',
-    checkboxActiveBorder: '#cc3333',
-    buttonBorder: '#c6c6c6',
-    buttonActiveBackground: '#222222',
-    buttonForeground: '#cccccc',
-    zoomIconFilter: 'invert(100%)',
-    gridColor: '#484848'
-};
-
-const darkGuiColors = {
-    'motion-primary': '#ff4c4c',
-    'motion-tertiary': '#cc3333',
-
-    'looks-secondary': '#ff4c4c',
-    'looks-transparent': '#ff4d4d59',
-    'looks-light-transparent': '#ff4d4d26',
-    'looks-secondary-dark': 'hsla(0, 42%, 51%, 1)',
-
-    'extensions-primary': 'hsla(10, 85%, 65%, 1)',
-    'extensions-tertiary': 'hsla(10, 85%, 40%, 1)',
-    'extensions-transparent': 'hsla(10, 85%, 65%, 0.35)',
-    'extensions-light': 'hsla(10, 57%, 85%, 1)',
-
-    'drop-highlight': '#ff8c8c'
-};
-
-export {
-    lightGuiColors,
-    darkBlockColors,
-    darkGuiColors
-};
